Collision edited and implemented with the stage

+++ core/src/com/mygdx/game/MovingObj.java
@@ -2,8 +2,6 @@ package com.mygdx.game;
 
 public abstract class MovingObj extends Object {
 
-    
-
     /**
      * 
      * @param posX
@@ -13,7 +11,7 @@ public abstract class MovingObj extends Object {
      * @param isCollidable
      * @param isVisible
      */
-    public MovingObj(int posX, int posY, int sizeX, int sizeY, boolean isCollidable, boolean isVisible) {
+    public MovingObj(float posX, float posY, float sizeX, float sizeY, boolean isCollidable, boolean isVisible) {
         super(posX, posY, sizeX, sizeY, isCollidable, isVisible);
     }
 }
